pyproject.toml
[project]
name = "mcp-mvp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.26.0",
    "yfinance>=0.2.0",
    "duckduckgo-search>=6.0.0",
    "psycopg2-binary>=2.9.0",
]
